+++ model_app/frontend/src/containers/TeamApplication.js
@@ -1,13 +1,6 @@
-import { React, useState } from 'react';
-import Button from '@material-ui/core/Button';
-import InputLabel from '@material-ui/core/InputLabel';
-import MenuItem from '@material-ui/core/MenuItem';
-import Select from '@material-ui/core/Select';
-import TextField from '@material-ui/core/TextField';
+import React from 'react';
 import Typography from '@material-ui/core/Typography';
-import Grid from '@material-ui/core/Grid';
 import Box from '@material-ui/core/Box';
-import StyledButton from '../components/StyledButton';
 import Card from '@material-ui/core/Card';
 import { Container } from '@material-ui/core';
 import { makeStyles } from '@material-ui/core/styles';
@@ -28,20 +21,9 @@ const useStyles = makeStyles((theme) => ({
 
 const TeamApplication = function TeamApplication() {
 
-  const [first_name, setFirst_name] = useState("");
-  const [last_name, setLast_name] = useState("");
-  const [major, setMajor] = useState("");
-
-  const [grad_year, setGrad_year] = useState("");
-
-  const [first_interested_team, setFirst_interested_team] = useState("");
-  const [second_interested_team, setSecond_interested_team] = useState("");
-  const [third_interested_team, setThird_interested_team] = useState("");
-  const [submitMsg, setSubmitMsg] = useState("")
   const classes = useStyles();
 
   return (
-    <Container style={{ padding: '30px 20%', backgroundColor: '#222629', margin: '0' }}>
       <Card style={{ justifyContent: 'center', padding: '30px 15%', backgroundColor: 'white' }}>
 
         <Box py={2} bgcolor="white" style={{ color: "#222629" }} align="center" >
@@ -50,104 +32,14 @@ const TeamApplication = function TeamApplication() {
 					</Typography>
 					<div className={classes.border}></div>
 
-          <form noValidate autoComplete="on">
-            <Grid container direction={"column"} spacing={3}>
-              <Grid item>
-                <Typography >
-                  First Name
-                </Typography>
-                <TextField required id="standard-required" variant="outlined" label="First Name" onChange={e => setFirst_name(e.target.value)} />
-              </Grid>
-              <Grid item>
-                <Typography >
-                  Last Name
-                </Typography>
-                <TextField required id="standard-basic" variant="outlined" label="Last Name" onChange={e => setLast_name(e.target.value)} />
-              </Grid>
 
-              <Grid item>
-                <Typography >
-                  Major
-                </Typography>
-                <TextField required id="standard-basic" variant="outlined" label="Major" onChange={e => setMajor(e.target.value)} />
-              </Grid>
-              <Grid item>
-                <Typography  >
-                  Graduation Year
-                </Typography>
-                <TextField required id="standard-basic" variant="outlined" label="Graduation Year" onChange={e => setGrad_year(e.target.value)} />
-              </Grid>
-
-              <Grid item>
-                <Typography>First Choice Team</Typography>
-                <Select
-                  onChange={(e) => {
-                    setFirst_interested_team(e.target.value);
-                  }}
-                  style={{ minWidth: "230px" }}
-                >
-                  <MenuItem value="Fullstack" style={{ backgroundColor: "white" }}>Fullstack</MenuItem>
-                  <MenuItem value="Simulation" style={{ backgroundColor: "white" }}>Simulation</MenuItem>
-                  <MenuItem value="Machine learning" style={{ backgroundColor: "white" }}>Machine learning</MenuItem>
-                  <MenuItem value="Information Support" style={{ backgroundColor: "white" }}>Information Support</MenuItem>
-                </Select>
-
-
-              </Grid>
-
-              <Grid item>
-                <Typography>Second Choice Team</Typography>
-                <Select
-                  onChange={(e) => {
-                    setSecond_interested_team(e.target.value);
-                  }}
-                  style={{ minWidth: "230px" }}
-                >
-                  <MenuItem value="Fullstack" style={{ backgroundColor: "white" }}>Fullstack</MenuItem>
-                  <MenuItem value="Simulation" style={{ backgroundColor: "white" }}>Simulation</MenuItem>
-                  <MenuItem value="Machine learning" style={{ backgroundColor: "white" }}>Machine learning</MenuItem>
-                  <MenuItem value="Information Support" style={{ backgroundColor: "white" }}>Information Support</MenuItem>
-                  <MenuItem value="None" style={{ backgroundColor: "white" }}>None</MenuItem>
-                </Select>
-
-
-              </Grid>
-
-              <Grid item>
-                <Typography>Third Choice Team</Typography>
-                <Select
-                  onChange={(e) => {
-                    setThird_interested_team(e.target.value);
-                  }}
-                  style={{ minWidth: "230px" }}
-                >
-                  <MenuItem value="Fullstack" style={{ backgroundColor: "white" }}>Fullstack</MenuItem>
-                  <MenuItem value="Simulation" style={{ backgroundColor: "white" }}>Simulation</MenuItem>
-                  <MenuItem value="Machine learning" style={{ backgroundColor: "white" }}>Machine learning</MenuItem>
-                  <MenuItem value="Information Support" style={{ backgroundColor: "white" }}>Information Support</MenuItem>
-                  <MenuItem value="None" style={{ backgroundColor: "white" }}>None</MenuItem>
-                </Select>
-
-
-              </Grid>
-
-              <Grid>
-                <StyledButton style={{ color: '#222629', marginTop: '30px' }}>
-                  Submit
-              </StyledButton>
-              </Grid>
-
-              <Grid>
-                <Typography style={{ color: "red" }}>
-                  {submitMsg}
-                </Typography>
-              </Grid>
-
-            </Grid >
-          </form>
+          <Typography variant="h5" component="h5" className={classes.title}>
+						If you are a Hopkins student and are interested in joining our team. </Typography>
+          <Typography variant="h5" component="h5" className={classes.title}>
+          Send us an email at delineodiseasemodeling@gmail.com with your resume and interested team(Fullstack, Simulation, Machine learning, Information Support)
+					</Typography>
         </Box>
       </Card>
-    </Container>
   )
 }
 export default TeamApplication;
